+++ morpha/morpha/models/__init__.py
@@ -2,10 +2,12 @@
 
 from .irecord import IRecord
 from .damage_record import DamageRecord
+from .move_record import MoveRecord
 from .team_preview_record import TeamPreviewRecord
 from .battle_log import BattleLog
 
 __all__ = ['BattleLog',
            'DamageRecord',
            'IRecord',
+           'MoveRecord',
            'TeamPreviewRecord']