+++ morpha/documentation/release_notes.md
@@ -4,4 +4,4 @@
 ##### Features
 - Added logging.
 - Added BattleMetricsService (`ME-630`).
-- Added BattleLog, IRecord, DamageRecord, and TeamPreviewRecord models (`ME-630`).
+- Added BattleLog, IRecord, DamageRecord, MoveRecord, and TeamPreviewRecord models (`ME-630`).